+++ Frends.Exchange.SendEmail/Frends.Exchange.SendEmail.Test/UnitTests.cs
@@ -14,6 +14,7 @@ public class UnitTests
     private readonly string? _password = Environment.GetEnvironmentVariable("Exchange_User_Password");
     private readonly string? _applicationID = Environment.GetEnvironmentVariable("Exchange_Application_ID");
     private readonly string? _tenantID = Environment.GetEnvironmentVariable("Exchange_Tenant_ID");
+    private static readonly string? _clientSecret = Environment.GetEnvironmentVariable("Exchange_ClientSecret");
     private static Connection _connection = new();
     private static Input _input = new();
     private static Options _options = new();
@@ -34,8 +35,8 @@ public class UnitTests
             Password = _password,
             ClientId = _applicationID,
             TenantId = _tenantID,
-            AuthenticationProvider = AuthenticationProviders.UsernamePassword,
-            ClientSecret = null,
+            AuthenticationProvider = AuthenticationProviders.ClientCredentialsSecret,
+            ClientSecret = _clientSecret,
             X509CertificateFilePath = null,
         };
 
@@ -56,7 +57,7 @@ public class UnitTests
         _options = new Options()
         {
             ThrowExceptionIfAttachmentNotFound = true,
-            ThrowExceptionOnFailure = true,
+            ThrowErrorOnFailure = true,
         };
     }
 
@@ -67,79 +68,79 @@ public class UnitTests
     }
 
     [TestMethod]
-    public async Task SendEmailTest_PlainText_UsernamePassword()
+    public async Task SendEmailTest_PlainText_ClientCredentialsSecret()
     {
-        _input.Subject = $"{_input.Subject}, Method: SendEmailTest_PlainText_UsernamePassword";
-        var result = await Exchange.SendEmail(_connection, _input, _options, default);
+        _input.Subject = $"{_input.Subject}, Method: SendEmailTest_PlainText_ClientCredentialsSecret";
+        var result = await Exchange.SendEmail(_input, _connection, _options, default);
         Assert.IsTrue(result.Success);
         Assert.AreEqual("Email sent successfully.", result.Data);
     }
 
     [TestMethod]
-    public async Task SendEmailTest_ToMultiple_Semicolon_UsernamePassword()
+    public async Task SendEmailTest_ToMultiple_Semicolon_ClientCredentialsSecret()
     {
-        _input.Subject = $"{_input.Subject}, Method: SendEmailTest_ToMultiple_Semicolon_UsernamePassword";
+        _input.Subject = $"{_input.Subject}, Method: SendEmailTest_ToMultiple_Semicolon_ClientCredentialsSecret";
         _input.To = $"{_user}; {_user2}";
-        var result = await Exchange.SendEmail(_connection, _input, _options, default);
+        var result = await Exchange.SendEmail(_input, _connection, _options, default);
         Assert.IsTrue(result.Success);
         Assert.AreEqual("Email sent successfully.", result.Data);
     }
 
     [TestMethod]
-    public async Task SendEmailTest_ToMultiple_Comma_UsernamePassword()
+    public async Task SendEmailTest_ToMultiple_Comma_ClientCredentialsSecret()
     {
-        _input.Subject = $"{_input.Subject}, Method: SendEmailTest_ToMultiple_Comma_UsernamePassword";
-        _input.To = $"{_user}; {_user2}";
-        var result = await Exchange.SendEmail(_connection, _input, _options, default);
+        _input.Subject = $"{_input.Subject}, Method: SendEmailTest_ToMultiple_Comma_ClientCredentialsSecret";
+        _input.To = $"{_user}, {_user2}";
+        var result = await Exchange.SendEmail(_input, _connection, _options, default);
         Assert.IsTrue(result.Success);
         Assert.AreEqual("Email sent successfully.", result.Data);
     }
 
     [TestMethod]
-    public async Task SendEmailTest_WithCC_UsernamePassword()
+    public async Task SendEmailTest_WithCC_ClientCredentialsSecret()
     {
-        _input.Subject = $"{_input.Subject}, Method: SendEmailTest_WithCC_UsernamePassword";
-        _input.Cc = $"{_user}; {_user2}";
-        var result = await Exchange.SendEmail(_connection, _input, _options, default);
+        _input.Subject = $"{_input.Subject}, Method: SendEmailTest_WithCC_ClientCredentialsSecret";
+        _input.Cc = $"{_user}, {_user2}";
+        var result = await Exchange.SendEmail(_input, _connection, _options, default);
         Assert.IsTrue(result.Success);
         Assert.AreEqual("Email sent successfully.", result.Data);
     }
 
     [TestMethod]
-    public async Task SendEmailTest_WithBCC_UsernamePassword()
+    public async Task SendEmailTest_WithBCC_ClientCredentialsSecret()
     {
-        _input.Subject = $"{_input.Subject}, Method: SendEmailTest_WithBCC_UsernamePassword";
-        _input.Bcc = $"{_user}; {_user2}";
-        var result = await Exchange.SendEmail(_connection, _input, _options, default);
+        _input.Subject = $"{_input.Subject}, Method: SendEmailTest_WithBCC_ClientCredentialsSecret";
+        _input.Bcc = $"{_user}, {_user2}";
+        var result = await Exchange.SendEmail(_input, _connection, _options, default);
         Assert.IsTrue(result.Success);
         Assert.AreEqual("Email sent successfully.", result.Data);
     }
 
     [TestMethod]
-    public async Task SendEmailTest_Html_UsernamePassword()
+    public async Task SendEmailTest_Html_ClientCredentialsSecret()
     {
-        _input.Subject = $"{_input.Subject}, Method: SendEmailTest_Html_UsernamePassword";
+        _input.Subject = $"{_input.Subject}, Method: SendEmailTest_Html_ClientCredentialsSecret";
         _input.Message = "<div><h1>This is a header text.</h1></div>";
         _input.IsMessageHtml = true;
-        var result = await Exchange.SendEmail(_connection, _input, _options, default);
+        var result = await Exchange.SendEmail(_input, _connection, _options, default);
         Assert.IsTrue(result.Success);
         Assert.AreEqual("Email sent successfully.", result.Data);
     }
 
     [TestMethod]
-    public async Task SendEmailTest_NordicLetters_UsernamePassword()
+    public async Task SendEmailTest_NordicLetters_ClientCredentialsSecret()
     {
-        _input.Subject = $"{_input.Subject}, Method: SendEmailTest_NordicLetters_UsernamePassword";
+        _input.Subject = $"{_input.Subject}, Method: SendEmailTest_NordicLetters_ClientCredentialsSecret";
         _input.Message = "Tämä testimaili tuo yöllä ålannista.";
-        var result = await Exchange.SendEmail(_connection, _input, _options, default);
+        var result = await Exchange.SendEmail(_input, _connection, _options, default);
         Assert.IsTrue(result.Success);
         Assert.AreEqual("Email sent successfully.", result.Data);
     }
 
     [TestMethod]
-    public async Task SendEmailTest_SingleFileAttachment_UsernamePassword()
+    public async Task SendEmailTest_SingleFileAttachment_ClientCredentialsSecret()
     {
-        _input.Subject = $"{_input.Subject}, Method: SendEmailTest_SingleFileAttachment_UsernamePassword";
+        _input.Subject = $"{_input.Subject}, Method: SendEmailTest_SingleFileAttachment_ClientCredentialsSecret";
         _input.Attachments = new[] {
             new Attachments()
             {
@@ -149,15 +150,15 @@ public class UnitTests
             }
         };
 
-        var result = await Exchange.SendEmail(_connection, _input, _options, default);
+        var result = await Exchange.SendEmail(_input, _connection, _options, default);
         Assert.IsTrue(result.Success);
         Assert.AreEqual($"Email sent successfully.", result.Data);
     }
 
     [TestMethod]
-    public async Task SendEmailTest_MultipleFileAttachment_UsernamePassword()
+    public async Task SendEmailTest_MultipleFileAttachment_ClientCredentialsSecret()
     {
-        _input.Subject = $"{_input.Subject}, Method: SendEmailTest_MultipleFileAttachment_UsernamePassword";
+        _input.Subject = $"{_input.Subject}, Method: SendEmailTest_MultipleFileAttachment_ClientCredentialsSecret";
         _input.Attachments = new[] {
             new Attachments()
             {
@@ -173,15 +174,15 @@ public class UnitTests
             }
         };
 
-        var result = await Exchange.SendEmail(_connection, _input, _options, default);
+        var result = await Exchange.SendEmail(_input, _connection, _options, default);
         Assert.IsTrue(result.Success);
         Assert.AreEqual($"Email sent successfully.", result.Data);
     }
 
     [TestMethod]
-    public async Task SendEmailTest_SingleFileFromDir_UsernamePassword()
+    public async Task SendEmailTest_SingleFileFromDir_ClientCredentialsSecret()
     {
-        _input.Subject = $"{_input.Subject}, Method: SendEmailTest_SingleFileFromDir_UsernamePassword";
+        _input.Subject = $"{_input.Subject}, Method: SendEmailTest_SingleFileFromDir_ClientCredentialsSecret";
         _input.Attachments = new[] {
             new Attachments()
             {
@@ -191,16 +192,15 @@ public class UnitTests
             }
         };
 
-        var result = await Exchange.SendEmail(_connection, _input, _options, default);
+        var result = await Exchange.SendEmail(_input, _connection, _options, default);
         Assert.IsTrue(result.Success);
         Assert.AreEqual($"Email sent successfully.", result.Data);
     }
 
     [TestMethod]
-    public async Task SendEmailTest_LargeAttachment_UsernamePassword()
+    public async Task SendEmailTest_LargeAttachment_ClientCredentialsSecret()
     {
-        _input.Subject = $"{_input.Subject}, Method: SendEmailTest_LargeAttachment_UsernamePassword";
-        _input.From = null;
+        _input.Subject = $"{_input.Subject}, Method: SendEmailTest_LargeAttachment_ClientCredentialsSecret";
         _input.Attachments = new[] {
             new Attachments()
             {
@@ -210,16 +210,15 @@ public class UnitTests
             }
         };
 
-        var result = await Exchange.SendEmail(_connection, _input, _options, default);
+        var result = await Exchange.SendEmail(_input, _connection, _options, default);
         Assert.IsTrue(result.Success);
         Assert.AreEqual($"Email sent successfully.", result.Data);
     }
 
     [TestMethod]
-    public async Task SendEmailTest_MultipleLargeAttachments_UsernamePassword()
+    public async Task SendEmailTest_MultipleLargeAttachments_ClientCredentialsSecret()
     {
-        _input.Subject = $"{_input.Subject}, Method: SendEmailTest_MultipleLargeAttachments_UsernamePassword";
-        _input.From = null;
+        _input.Subject = $"{_input.Subject}, Method: SendEmailTest_MultipleLargeAttachments_ClientCredentialsSecret";
         _input.Attachments = new[] {
             new Attachments()
             {
@@ -235,16 +234,16 @@ public class UnitTests
             }
         };
 
-        var result = await Exchange.SendEmail(_connection, _input, _options, default);
+        var result = await Exchange.SendEmail(_input, _connection, _options, default);
         Assert.IsTrue(result.Success);
         Assert.AreEqual($"Email sent successfully.", result.Data);
     }
 
     [TestMethod]
-    public async Task SendEmailTest_StringAttachment_UsernamePassword()
+    public async Task SendEmailTest_StringAttachment_ClientCredentialsSecret()
     {
 
-        _input.Subject = $"{_input.Subject}, Method: SendEmailTest_StringAttachment_UsernamePassword";
+        _input.Subject = $"{_input.Subject}, Method: SendEmailTest_StringAttachment_ClientCredentialsSecret";
         _input.Attachments = new[] {
             new Attachments()
             {
@@ -254,16 +253,16 @@ public class UnitTests
             }
         };
 
-        var result = await Exchange.SendEmail(_connection, _input, _options, default);
+        var result = await Exchange.SendEmail(_input, _connection, _options, default);
         Assert.IsTrue(result.Success);
         Assert.AreEqual($"Email sent successfully.", result.Data);
     }
 
     [TestMethod]
-    public async Task SendEmailTest_MultipleStringAttachment_UsernamePassword()
+    public async Task SendEmailTest_MultipleStringAttachment_ClientCredentialsSecret()
     {
 
-        _input.Subject = $"{_input.Subject}, Method: SendEmailTest_MultipleStringAttachment_UsernamePassword";
+        _input.Subject = $"{_input.Subject}, Method: SendEmailTest_MultipleStringAttachment_ClientCredentialsSecret";
         _input.Attachments = new[] {
             new Attachments()
             {
@@ -279,16 +278,16 @@ public class UnitTests
             }
         };
 
-        var result = await Exchange.SendEmail(_connection, _input, _options, default);
+        var result = await Exchange.SendEmail(_input, _connection, _options, default);
         Assert.IsTrue(result.Success);
         Assert.AreEqual($"Email sent successfully.", result.Data);
     }
 
     [TestMethod]
-    public async Task SendEmailTest_MixAttachments_UsernamePassword()
+    public async Task SendEmailTest_MixAttachments_ClientCredentialsSecret()
     {
 
-        _input.Subject = $"{_input.Subject}, Method: SendEmailTest_MixAttachments_UsernamePassword";
+        _input.Subject = $"{_input.Subject}, Method: SendEmailTest_MixAttachments_ClientCredentialsSecret";
         _input.Attachments = new[] {
             new Attachments()
             {
@@ -304,7 +303,7 @@ public class UnitTests
             }
         };
 
-        var result = await Exchange.SendEmail(_connection, _input, _options, default);
+        var result = await Exchange.SendEmail(_input, _connection, _options, default);
         Assert.IsTrue(result.Success);
         Assert.AreEqual($"Email sent successfully.", result.Data);
     }
@@ -318,7 +317,7 @@ public class UnitTests
     {
         _input.From = _user2;
         _input.Subject = $"{_input.Subject}, Method: SendEmailTest_SendEmailAsAnotherUser_UsernamePassword";
-        var result = await Exchange.SendEmail(_connection, _input, _options, default);
+        var result = await Exchange.SendEmail(_input, _connection, _options, default);
         Assert.IsTrue(result.Success);
     }
 
@@ -326,26 +325,28 @@ public class UnitTests
     public async Task SendEmailTest_MissingTo_Throw()
     {
         _input.To = null;
-        await Assert.ThrowsExceptionAsync<ArgumentNullException>(async () => await Exchange.SendEmail(_connection, _input, _options, default));
+        await Assert.ThrowsExceptionAsync<ArgumentNullException>(async () => await Exchange.SendEmail(_input, _connection, _options, default));
     }
 
     [TestMethod]
     public async Task SendEmailTest_MissingCredentials_UsernamePassword_Throw()
     {
+        _connection.AuthenticationProvider = AuthenticationProviders.UsernamePassword;
+
         _connection.TenantId = null;
-        await Assert.ThrowsExceptionAsync<ArgumentNullException>(async () => await Exchange.SendEmail(_connection, _input, _options, default));
+        await Assert.ThrowsExceptionAsync<ArgumentNullException>(async () => await Exchange.SendEmail(_input, _connection, _options, default));
 
         _connection.TenantId = _tenantID;
         _connection.ClientId = null;
-        await Assert.ThrowsExceptionAsync<ArgumentNullException>(async () => await Exchange.SendEmail(_connection, _input, _options, default));
+        await Assert.ThrowsExceptionAsync<ArgumentNullException>(async () => await Exchange.SendEmail(_input, _connection, _options, default));
 
         _connection.ClientId = _applicationID;
         _connection.Username = null;
-        await Assert.ThrowsExceptionAsync<ArgumentNullException>(async () => await Exchange.SendEmail(_connection, _input, _options, default));
+        await Assert.ThrowsExceptionAsync<ArgumentNullException>(async () => await Exchange.SendEmail(_input, _connection, _options, default));
 
         _connection.Username = _user;
         _connection.Password = null;
-        await Assert.ThrowsExceptionAsync<ArgumentNullException>(async () => await Exchange.SendEmail(_connection, _input, _options, default));
+        await Assert.ThrowsExceptionAsync<ArgumentNullException>(async () => await Exchange.SendEmail(_input, _connection, _options, default));
     }
 
     [TestMethod]
@@ -355,15 +356,15 @@ public class UnitTests
 
         _connection.X509CertificateFilePath = "Something";
         _connection.TenantId = null;
-        await Assert.ThrowsExceptionAsync<ArgumentNullException>(async () => await Exchange.SendEmail(_connection, _input, _options, default));
+        await Assert.ThrowsExceptionAsync<ArgumentNullException>(async () => await Exchange.SendEmail(_input, _connection, _options, default));
 
         _connection.TenantId = _tenantID;
         _connection.ClientId = null;
-        await Assert.ThrowsExceptionAsync<ArgumentNullException>(async () => await Exchange.SendEmail(_connection, _input, _options, default));
+        await Assert.ThrowsExceptionAsync<ArgumentNullException>(async () => await Exchange.SendEmail(_input, _connection, _options, default));
 
         _connection.ClientId = _applicationID;
         _connection.X509CertificateFilePath = null;
-        await Assert.ThrowsExceptionAsync<ArgumentNullException>(async () => await Exchange.SendEmail(_connection, _input, _options, default));
+        await Assert.ThrowsExceptionAsync<ArgumentNullException>(async () => await Exchange.SendEmail(_input, _connection, _options, default));
     }
 
     [TestMethod]
@@ -373,15 +374,15 @@ public class UnitTests
 
         _connection.ClientSecret = "Something";
         _connection.TenantId = null;
-        await Assert.ThrowsExceptionAsync<ArgumentNullException>(async () => await Exchange.SendEmail(_connection, _input, _options, default));
+        await Assert.ThrowsExceptionAsync<ArgumentNullException>(async () => await Exchange.SendEmail(_input, _connection, _options, default));
 
         _connection.TenantId = _tenantID;
         _connection.ClientId = null;
-        await Assert.ThrowsExceptionAsync<ArgumentNullException>(async () => await Exchange.SendEmail(_connection, _input, _options, default));
+        await Assert.ThrowsExceptionAsync<ArgumentNullException>(async () => await Exchange.SendEmail(_input, _connection, _options, default));
 
         _connection.ClientId = _applicationID;
         _connection.ClientSecret = null;
-        await Assert.ThrowsExceptionAsync<ArgumentNullException>(async () => await Exchange.SendEmail(_connection, _input, _options, default));
+        await Assert.ThrowsExceptionAsync<ArgumentNullException>(async () => await Exchange.SendEmail(_input, _connection, _options, default));
     }
 
     [TestMethod]
@@ -396,12 +397,14 @@ public class UnitTests
                 FileMask = "*.*"
             }
         };
-        _options.ThrowExceptionOnFailure = false;
+        _options.ThrowErrorOnFailure = false;
         _options.ThrowExceptionIfAttachmentNotFound = true;
 
-        var result = await Exchange.SendEmail(_connection, _input, _options, default);
+        var result = await Exchange.SendEmail(_input, _connection, _options, default);
         Assert.IsFalse(result.Success);
         Assert.AreEqual("Failed to send an email. No files found in directory C:\\nothinghere.", result.Data);
+        Assert.IsNotNull(result.Error);
+        Assert.AreEqual("No files found in directory C:\\nothinghere.", result.Error.Message);
     }
 
     [TestMethod]
@@ -416,9 +419,9 @@ public class UnitTests
                 FileMask = "*.*"
             }
         };
-        _options.ThrowExceptionOnFailure = true;
+        _options.ThrowErrorOnFailure = true;
         _options.ThrowExceptionIfAttachmentNotFound = true;
-        await Assert.ThrowsExceptionAsync<Exception>(async () => await Exchange.SendEmail(_connection, _input, _options, default));
+        await Assert.ThrowsExceptionAsync<Exception>(async () => await Exchange.SendEmail(_input, _connection, _options, default));
     }
 
     internal static void CreateFiles()