+++ Frends.Exchange.SendEmail/Frends.Exchange.SendEmail/Definitions/Attachments.cs
@@ -9,7 +9,7 @@ namespace Frends.Exchange.SendEmail.Definitions;
 public class Attachments
 {
     /// <summary>
-    /// Specifies whether the attachment file is created from a string or copied from disk. 
+    /// Specifies whether the attachment file is created from a string or copied from disk.
     /// This determines how the attachment is added to the email.
     /// </summary>
     /// <example>AttachmentTypes.FileAttachment</example>
@@ -24,14 +24,14 @@ public class Attachments
     public string FileName { get; set; }
 
     /// <summary>
-    /// The content of the attachment file. 
+    /// The content of the attachment file.
     /// </summary>
     /// <example>Hello, World!</example>
     [UIHint(nameof(AttachmentType), "", AttachmentTypes.AttachmentFromString)]
     public string FileContent { get; set; }
 
     /// <summary>
-    /// The path of the attachment file. 
+    /// The path of the attachment file.
     /// If the path ends in a directory, all files in that folder with the given Attachments.FileMask are added as attachments.
     /// </summary>
     /// <example>C:\temp\message.txt</example>
@@ -39,7 +39,7 @@ public class Attachments
     public string FilePath { get; set; }
 
     /// <summary>
-    /// The file mask. 
+    /// The file mask.
     /// </summary>
     /// <example>*.txt</example>
     [UIHint(nameof(AttachmentType), "", AttachmentTypes.FileAttachment)]