+++ Frends.Exchange.SendEmail/Frends.Exchange.SendEmail/Definitions/Result.cs
@@ -5,6 +5,13 @@
 /// </summary>
 public class Result
 {
+    internal Result(bool success, string data, Error error = null)
+    {
+        Success = success;
+        Data = data;
+        Error = error;
+    }
+
     /// <summary>
     /// Gets a value indicating whether the task was executed successfully.
     /// </summary>
@@ -12,14 +19,14 @@ public class Result
     public bool Success { get; private set; }
 
     /// <summary>
-    /// Gets the result of the task. Contains exception message if exception was thrown and Options.ThrowExceptionOnFailure = false.
+    /// Gets the result of the task. Contains exception message if exception was thrown and Options.ThrowErrorOnFailure = false.
     /// </summary>
     /// <example>Email sent successfully.</example>
     public string Data { get; private set; }
 
-    internal Result(bool success, string data)
-    {
-        Success = success;
-        Data = data;
-    }
+    /// <summary>
+    /// Gets the error details of the task. Null when Success is true.
+    /// </summary>
+    /// <example>null</example>
+    public Error Error { get; private set; }
 }